feat(ui) : 코드리뷰를 참고받아 수정완료
